### 2025-09-24 #9073 @editorialbot
[PRE REVIEW]: mdopt: A code-agnostic tensor-network decoder for quantum error-correcting codes
**Submitting author:** @meandmytram (<a href="http://orcid.org/0009-0002-8149-135X">Aleksandr Berezutskii</a>)
**Repository:** https://github.com/quicophy/mdopt
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
