+++ paper/paper.bib
@@ -6,14 +6,16 @@
   number={3},
   pages={030501},
   year={2014},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevLett.113.030501}
 }
 
 @article{chubb2021general,
   title={General tensor network decoding of 2D Pauli codes},
   author={Chubb, Christopher T},
   journal={arXiv preprint arXiv:2101.04125},
-  year={2021}
+  year={2021},
+  doi={10.48550/arXiv.2101.04125}
 }
 
 @article{piveteau2024tensor,
@@ -24,28 +26,32 @@
   number={4},
   pages={040303},
   year={2024},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PRXQuantum.5.040303}
 }
 
 @article{higgott2021pymatching,
   title={PyMatching: A Python package for decoding quantum codes with minimum-weight perfect matching, arXiv e-prints},
   author={Higgott, Oscar},
   journal={arXiv preprint arXiv:2105.13082},
-  year={2021}
+  year={2021},
+  doi={10.48550/arXiv.2105.13082}
 }
 
 @phdthesis{qecsim,
-  author = {Tuckett, David Kingsley},
-  title = {Tailoring surface codes: Improvements in quantum error correction with biased noise},
-  school = {University of Sydney},
-  doi = {10.25910/x8xw-9077},
-  year = {2020},
-  note = {(qecsim: \url{https://github.com/qecsim/qecsim})}
+  author={Tuckett, David Kingsley},
+  title={Tailoring surface codes: Improvements in quantum error correction with biased noise},
+  school={University of Sydney},
+  doi={10.25910/x8xw-9077},
+  year={2020},
+  note={(qecsim: \url{https://github.com/qecsim/qecsim})},
+  doi={10.25910/x8xw-9077},
 }
 
 @article{gidney2021stim,
   title={Stim: a fast stabilizer circuit simulator},
   author={Gidney, Craig},
   journal={arXiv preprint arXiv:2103.02202},
-  year={2021}
+  year={2021},
+  doi={10.22331/q-2021-07-06-497}
 }
