+++ paper/paper.md
@@ -19,13 +19,13 @@ bibliography: paper.bib
 
 # Summary
 
-`mdopt` is an open-source Python library for code-agnostic decoding of quantum error-correcting codes using tensor networks (in particular, MPS and MPO). Given only the parity-check matrices (e.g., \(H_X, H_Z\)) and a noise model, `mdopt` builds a tensor-network representation of the decoding problem in the form of MPS-MPO evolution and contracts it to approximate (or, for small instances, exact) maximum-likelihood posteriors over logical operators. Depending on the number of logical qubits, the code either reads out the posterior maximum directly or uses a DMRG-like variational algorithm to find the most likely logical correction. The library targets researchers who wish to benchmark codes and noise models beyond simple settings, while retaining a clear and reproducible workflow in pure Python.
+`mdopt` is an open-source Python library for code-agnostic decoding of quantum error-correcting codes using tensor networks (in particular, MPS and MPO). Given only the parity-check matrices and a noise model, `mdopt` builds a tensor-network representation of the decoding problem in the form of MPS-MPO evolution and contracts it to approximate (or, for small instances, exact) maximum-likelihood posteriors over logical operators. Depending on the number of logical qubits, the code either reads out the posterior maximum directly or uses a DMRG-like variational algorithm to find the most likely logical correction. The library targets researchers who wish to benchmark codes and noise models beyond simple settings, while retaining a clear and reproducible workflow in pure Python.
 
 # Statement of need
 
 Decoding is central to assessing the performance of quantum codes and hardware. Many widely used decoders are tailored to specific code families, e.g., Minimum Weight Perfect Matching (MWPM) for matching-like detection graphs. While these tools are powerful, researchers frequently need a generic decoder to compare heterogeneous codes and noise models under a common, transparent methodology. Tensor-network methods provide such perspective: decoding can be cast as contracting a TN derived from the code to compute likelihoods of logical events given a syndrome [@ferris2014tensor]. TN decoders have achieved state-of-the-art accuracy for 2D codes [@chubb2021general] and have recently been extended beyond 2D and to circuit-level noise [@piveteau2024tensor]. However, prior TN decoders primarily target codes with just a few logical qubits; open-source tools for multi-logical quantum LDPC (qLDPC) codes have been lacking. `mdopt` addresses this gap with a DMRG-like variational sweep that directly reads out the most likely logical configuration across multiple logical qubits, providing a novel, code-agnostic TN decoder for qLDPC codes, while remaining lightweight and reproducible.
 
-`mdopt` fills a gap by providing an open-source, Python-first implementation of a code-agnostic TN decoder:
+`mdopt` fills the gap by providing an open-source, Python-first implementation of a code-agnostic TN decoder:
 
 - **Inputs:** parity-check matrices, a description of logical observables, and a parametric noise model.
 - **Outputs:** logical-error posteriors and sample-level decisions (success/failure), with utilities to aggregate failure-rate curves.
