+++ paper/paper.bib
@@ -51,7 +51,10 @@
 @article{gidney2021stim,
   title={Stim: a fast stabilizer circuit simulator},
   author={Gidney, Craig},
-  journal={arXiv preprint arXiv:2103.02202},
+  journal={Quantum},
+  volume={5},
+  pages={497},
   year={2021},
+  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften},
   doi={10.22331/q-2021-07-06-497}
 }
