### 2025-09-26 #9073 @meandmytram
Hi @danielskatz,

Here are my suggestions: Roger Luo, Hector Mozo, Akash Kant, Mehmet Kececi, Ema Puljak, Lukas Burgholzer, Zhen Huang, Martin Beseda.

As for my previous suggestions, you can contact them in the following manner.
Ema Puljak: https://www.emapuljak.com/
Peter-Jan Derks: https://www.physik.fu-berlin.de/en/einrichtungen/ag/ag-eisert/team/phd/derks/index.html

Thanks,
Aleksandr
