### 2025-09-29 #9073 @danielskatz
👋 @emapuljak & @peter-janderks, I just wanted to check with you again on this, based on the author's suggestion. Would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
