### 2025-09-29 #9073 @burgholzer
@danielskatz I could find some time to review this, although quantum error correction is not explicitly my area of expertise. If that's okay, I am happy to focus my review primarily on the coding/packaging aspects.
