### 2025-09-30 #9073 @WhiteSymmetry
Python 3.13 for Windows:

pip install mdopt

\Lib\site-packages\setuptools\version.py", line 1, in <module>
          import pkg_resources
        File "\Lib\site-packages\pkg_resources\__init__.py", line 2172, in <module>
          register_finder(pkgutil.ImpImporter, find_on_path)
                          ^^^^^^^^^^^^^^^^^^^
      AttributeError: module 'pkgutil' has no attribute 'ImpImporter'. Did you mean: 'zipimporter'?
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

pip install git+https://github.com/quicophy/mdopt.git

p\pip-install-fp3xe5tp\numpy_80f267bb442a410b896cd4858a85b7cf\.mesonpy-cg84vy3u
      Build type: native build
      Project name: NumPy
      Project version: 1.26.4

      ..\meson.build:1:0: ERROR: Compiler cl cannot compile programs.

      A full log can be found at \AppData\Local\Temp\pip-install-fp3xe5tp\numpy_80f267bb442a410b896cd4858a85b7cf\.mesonpy-cg84vy3u\meson-logs\meson-log.txt
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.

Python 3.11 for Windows:
pip install git+https://github.com/quicophy/mdopt.git

Successfully installed diskcache-5.6.3 galois-0.4.6 ldpc-2.3.9 matrex-0.0.1 mdopt-0.5.0 numpy-1.26.4 qecsim-1.0b9 qecstruct-0.2.9 qldpc-0.1.4 sinter-1.15.0

File \Lib\site-packages\qldpc\codes\distance.py:8
      5 import numpy as np
      6 import numpy.typing as npt
----> 8 _mask55 = np.uint(0x5555555555555555)
      9 _mask33 = np.uint(0x3333333333333333)
     10 _mask0F = np.uint(0x0F0F0F0F0F0F0F0F)

1. qLDPC 0.1.4 --> qLDPC 0.2.4
2. (Numpy very old.) numpy 1.26.4 --> 2.3.3 or pyproject.toml: numpy = "^1.26.4" --> numpy = ">=1.26.4"
3. # renormalise=True, # TypeError: apply_bitflip_bias() got an unexpected keyword argument 'renormalise'
4. # result_to_explicit=False, # TypeError: apply_constraints() got an unexpected keyword argument 'result_to_explicit'
Code is working

"The code doesn't have a current pip installation setup. I think a current pip installation should be implemented to resolve errors and 'unexpected keyword argument' issues. After these are fixed, the code works - these errors don't occur in the examples within the 'examples' directory."
