+++ paper/paper.md
@@ -68,6 +68,6 @@ logicals, success = decode_css(
 
 # Acknowledgements
 
-This work was supported by the Ministère de l’Économie, de l’Innovation et de l’Énergie du Québec through its Research Chair in Quantum Computing, an NSERC Discovery grant, and the Canada First Research Excellence Fund. This work made use of the compute infrastructure of Calcul Québec and the Digital Research Alliance of Canada.
+This work was supported by the Ministère de l’Économie, de l’Innovation et de l’Énergie du Québec through its Research Chair in Quantum Computing, an NSERC Discovery grant, and the Canada First Research Excellence Fund. This work made use of the compute infrastructure of Calcul Québec and the Digital Research Alliance of Canada. A.B. thanks Stefanos Kourris for helpful discussions as well as the Unitary Foundation for supporting the project in the early stages.
 
 # References
