### 2025-10-01 #9073 @meandmytram
@WhiteSymmetry Thanks for the message. I have tried pip installation in a new environment on mac and it works. Unfortunately, I am too busy at the moment to dig into the Windows issue you are having so I will have to state explicitly that the package won't work with Windows, sorry.
