### 2025-10-02 quicophy/mdopt#419 @HectorMozo3110
Suggestions for Documentation (JOSS review)
👋 @meandmytram

After reviewing the current documentation, I think it provides a good starting point (installation instructions, examples, API reference). However, for new users the flow may feel a bit fragmented and difficult to follow. The order of sections is somewhat technical-first and could be reorganized to make the software easier to approach.

#### 1.Main observations
- The documentation starts abruptly with installation, without first giving context about the scientific problem solved.
- The sidebar highlights submodules early (`mdopt.contractor`, `mdopt.mps`), which is useful for advanced users, but intimidating for newcomers.
- Examples are present but listed without a clear pedagogical order (easier to harder).
- There is no **Quickstart** section where a user can simply copy-paste 5 lines of code and verify that the package works.
- The narrative lacks a clear "story" that guides the reader from installation → first run → advanced cases → API reference.

#### 2. Suggested structure (reorganized index)
1. **Introduction & Motivation**
   - What problem does `mdopt` solve?
   - In which research contexts is it relevant (e.g., tensor networks, quantum error correction)?

2. **Installation**
   - Keep `pip install mdopt` and `poetry install`

3. **Quickstart Example**
   - Minimal 5–10 line script showing a working optimisation

4. **Tutorials / Examples**
   - Order them from simple → complex
   - Add short narrative before each example (not only code)

5. **API Reference**
   - `mdopt.contractor`, `mdopt.mps`, `mdopt.optimiser`, `mdopt.utils`

6. **Contributing Guide**
   - How to run tests, code style, developer notes

7. **Declaration of Intended Use**
   - Who are the primary users? (physicists, quantum information researchers, optimisation community)
   - Typical use cases and expected outcomes

#### 3. Templates and references
To simplify improvements, it may help to look at documentation templates that are widely used in scientific software projects. For example:

- [Scientific Python Cookiecutter](https://github.com/NSLS-II/scientific-python-cookiecutter)
- [Cookiecutter Data Science](https://cookiecutter-data-science.drivendata.org/)
- [Software Sustainability Institute – Best Practices for Documentation](https://www.software.ac.uk/blog/what-are-best-practices-research-software-documentation)

You can either **adapt the current documentation** following this suggested index or **use a template from the projects above as a reference** to reorganize it more effectively.

---

👉 We can use this thread to discuss all documentation-related points. Once these suggestions are addressed and complemented, I will consider this thread closed.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/9125)
