### 2025-10-02 quicophy/mdopt#420 @HectorMozo3110
Observations on the README (JOSS review)
👋 @meandmytram

The current README provides a good foundation: it includes badges, installation instructions, a complete example, citation details, and contribution guidelines. However, several important elements are missing or could be improved to better align with JOSS expectations and to make the documentation more accessible for new users.

#### 1. Quickstart Example
The only example included is long and technical. Adding a short Quickstart (5–10 lines of code) would allow users to immediately test that the package works without needing to understand the full decoding workflow.

#### 2. Purpose and Intended Audience
The README does not explicitly state what kind of problems `mdopt` solves or who the expected users are. A short paragraph describing the scientific scope (tensor networks, quantum error correction, discrete optimisation) and the target community (physicists, quantum information researchers, optimisation specialists) would improve clarity.

#### 3. Examples Overview
Currently, the README only links to the `examples/` folder. It would be more user-friendly to briefly describe a few key examples (e.g., Ising model, LDPC codes, random circuits) so that newcomers know where to begin.

#### 4. Dependencies and Requirements
The README does not list dependencies or supported environments. Including the minimum Python version and core packages (`numpy`, `qecstruct`, etc.), as well as notes on supported platforms, would help users avoid installation issues.

#### 5. Developer Guide
While there are contribution guidelines, the README could provide more detail for developers:
- How to run the test suite (`pytest`, `tox`, etc.).
- How to build the documentation.
- Coding style and pre-commit conventions.

#### 6. License Section
Although the MIT license is mentioned via a badge, it would be clearer to include a short section explicitly linking to the `LICENSE` file.

#### 7. Link to Full Documentation
Although the README includes a ReadTheDocs badge, it does not explicitly direct the user to the full documentation. A dedicated section or line with a clear link to the online docs would be essential.
If desired, we could also refer back to the observations made in the other issue about documentation structure, since the README and the full docs should complement each other.

---

We can use this thread to discuss all README-related aspects. Once these points are addressed and complemented, I will consider this item closed.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/9125)
