### 2025-10-02 #9125 @HectorMozo3110
👋 @meandmytram

These are my initial observations based on the JOSS review checklist. As part of this, I opened a few issues that you’ve probably already noticed, mainly to make observations and suggestions for improvement.

We can use those threads to develop the discussion point by point. As we go covering them, I’ll check them off here. I will also update this list if I open additional issues, and once we’ve gone through all of them I’ll be able to close my review.

---
- [ ] [#419 Suggestions for Documentation (JOSS review)](https://github.com/quicophy/mdopt/issues/419)
- [ ] [#420 Observations on the README (JOSS review)](https://github.com/quicophy/mdopt/issues/420)
