+++ README.md
@@ -10,7 +10,7 @@
 [![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://lbesson.mit-license.org/)
 
 
-### `mdopt` is a python package built on top of numpy for discrete optimisation in the tensor-network (specifically, MPS-MPO) language.
+### `mdopt` is a python package built on top of `numpy` for discrete optimisation (mainly for classical and quantum decoding) in the tensor-network (specifically, Matrix Product States / Operators) language. The intended audience includes physicists, quantum information / error correction researchers, and those interested in exploring tensor-network methods beyond traditional condensed-matter applications.
 
 ## Installation
 
@@ -26,128 +26,44 @@ Otherwise, clone the repository and use [poetry](https://python-poetry.org/).
 poetry install
 ```
 
-## `mdopt` at a glance
+## Minimal example
 
 ```python
-import logging
 import numpy as np
 import qecstruct as qec
-from mdopt.mps.utils import create_custom_product_state
-from mdopt.optimiser.utils import (
-    SWAP,
-    XOR_BULK,
-    XOR_LEFT,
-    XOR_RIGHT,
-)
-from examples.decoding.decoding import (
-    linear_code_constraint_sites,
-    linear_code_prepare_message,
-)
-from examples.decoding.decoding import (
-    apply_bitflip_bias,
-    apply_constraints,
-    decode_message,
+from examples.decoding.decoding import decode_css
+
+# Define a small instance of the surface code
+LATTICE_SIZE = 3
+surface_code = qec.hypergraph_product(
+    qec.repetition_code(LATTICE_SIZE),
+    qec.repetition_code(LATTICE_SIZE),
 )
 
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+# Input an error and choose decoder controls
+logicals, success = decode_css(
+    code=surface_code,
+    error="IIXIIIIIIIIII",
+    bias_prob=0.1,
+    bias_type="Bitflip",
+    chi_max=64,
+    renormalise=True,
+    contraction_strategy="Optimised",
+    tolerance=0,
+    silent=False,
 )
-
-NUM_BITS = 24
-CHI_MAX = 256
-NUM_EXPERIMENTS = 10
-
-SEED = 123
-seed_seq = np.random.SeedSequence(SEED)
-
-error_rates = np.linspace(0.1, 0.3, 10)
-failures_statistics = {}
-
-for ERROR_RATE in error_rates:
-    logging.info(
-        f"Starting experiments for NUM_BITS={NUM_BITS}, CHI_MAX={CHI_MAX}, ERROR_RATE={ERROR_RATE}"
-    )
-    failures = []
-
-    for l in range(NUM_EXPERIMENTS):
-        new_seed = seed_seq.spawn(1)[0]
-        rng = np.random.default_rng(new_seed)
-        random_integer = rng.integers(1, 10**8 + 1)
-        SEED = random_integer
-
-        CHECK_DEGREE, BIT_DEGREE = 4, 3
-        NUM_CHECKS = int(BIT_DEGREE * NUM_BITS / CHECK_DEGREE)
-        if NUM_BITS / NUM_CHECKS != CHECK_DEGREE / BIT_DEGREE:
-            raise ValueError("The Tanner graph of the code must be bipartite.")
-        PROB_BIAS = ERROR_RATE
-
-        code = qec.random_regular_code(
-            NUM_BITS, NUM_CHECKS, BIT_DEGREE, CHECK_DEGREE, qec.Rng(SEED)
-        )
-        code_constraint_sites = linear_code_constraint_sites(code)
-
[…]
