### 2025-10-03 quicophy/mdopt#420 @meandmytram
hey @HectorMozo3110, thanks a lot for giving this a thorough read and making the comments, here are my responses point-by-point:

1. Addressed, the new shorter example now sits in the readme.
2. Addressed, I have added a short paragraph at the very beginning which talks briefly about these points.
3. Addressed, I have added a small piece of text describing the main examples.
4. I think this information can be easily inferred from the environment files in the repository. I have just added the statement that we do not support Windows for installation.
5. Addressed
6. Addressed
7. Addressed
