+++ docs/source/README.rst
@@ -4,8 +4,8 @@
 |codecov| |tests| |Documentation Status| |pre-commit.ci status| |lint|
 |mypy| |Unitary Fund| |MIT license|
 
-``mdopt`` is a python package built on top of numpy for discrete optimisation in the tensor-network (specifically, MPS-MPO) language.
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+``mdopt`` is a python package built on top of ``numpy`` for discrete optimisation (mainly for classical and quantum decoding) in the tensor-network (specifically, Matrix Product States / Operators) language. The intended audience includes physicists, quantum information / error correction researchers, and those interested in exploring tensor-network methods beyond traditional applications.
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Installation
 ------------
@@ -24,129 +24,51 @@ Otherwise, clone the repository and use
 
    poetry install
 
-``mdopt`` at a glance
----------------------
+Minimal example
+---------------
 
 .. code:: python
 
-   import logging
    import numpy as np
    import qecstruct as qec
-   from mdopt.mps.utils import create_custom_product_state
-   from mdopt.optimiser.utils import (
-       SWAP,
-       XOR_BULK,
-       XOR_LEFT,
-       XOR_RIGHT,
-   )
-   from examples.decoding.decoding import (
-       linear_code_constraint_sites,
-       linear_code_prepare_message,
-   )
-   from examples.decoding.decoding import (
-       apply_bitflip_bias,
-       apply_constraints,
-       decode_message,
+   from examples.decoding.decoding import decode_css
+
+   # Define a small instance of the surface code
+   LATTICE_SIZE = 3
+   surface_code = qec.hypergraph_product(
+       qec.repetition_code(LATTICE_SIZE),
+       qec.repetition_code(LATTICE_SIZE),
    )
 
-   logging.basicConfig(
-       level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+   # Input an error and choose decoder controls
+   logicals, success = decode_css(
+       code=surface_code,
+       error="IIXIIIIIIIIII",
+       bias_prob=0.1,
+       bias_type="Bitflip",
+       chi_max=64,
+       renormalise=True,
+       contraction_strategy="Optimised",
+       tolerance=0,
+       silent=False,
    )
 
-   NUM_BITS = 24
-   CHI_MAX = 256
-   NUM_EXPERIMENTS = 10
+Examples
+--------
 
-   SEED = 123
-   seed_seq = np.random.SeedSequence(SEED)
[…]
