+++ docs/source/README.rst
@@ -17,7 +17,7 @@ To install the current release, use the package manager
 
    pip install mdopt
 
-Otherwise, clone the repository and use
+Otherwise, you can clone the repository and use
 `poetry <https://python-poetry.org/>`__.
 
 .. code:: bash
@@ -57,18 +57,12 @@ Examples
 --------
 
 The `examples <https://github.com/quicophy/mdopt/tree/main/examples>`__
-folder contains full workflows that demonstrate typical use cases:
-
-- **Classical / quantum LDPC code decoding**: tensor-network decoder for
-  classical and quantum LDPC codes under various noise models.
-- **Ising model**: optimisation of spin Hamiltonians using MPS
-  ground-state search.
-- **Random circuits**: contraction of random tensor networks for
-  benchmarking.
-
-Each example is fully documented and serves as a starting point for
-building your own experiments. The package has been tested on macOS and
-Linux (Compute Canada clusters) and does not currently support Windows.
+folder contains full workflows that demonstrate typical use cases, such
+as quantum / classical LDPC code decoding, ground state search for the
+quantum Ising model and random quantum curcuit simulation. Each example
+is fully documented and serves as a starting point for building your own
+experiments. The package has been tested on macOS and Linux (Compute
+Canada clusters) and does not currently support Windows.
 
 Cite
 ----
@@ -129,8 +123,9 @@ Developers may find the following guidelines useful:
 
      black .
 
-- **Pre-commit hooks** Pre-commit hooks are configured to enforce
-  consistent style automatically. To enable them:
+- **Pre-commit hooks** `Pre-commit <https://pre-commit.com/>`__ hooks
+  are configured to enforce consistent style automatically. To enable
+  them:
 
   .. code:: bash
 
