### 2025-10-04 58ae3c4
update README.rst
