### 2025-10-04 6bfe44c
update README.rst
