+++ README.md
@@ -89,13 +89,13 @@ open-source software development, such as:
 
 Developers may find the following guidelines useful:
 
-- **Running tests**
-  Tests are executed using [pytest](https://docs.pytest.org/)
+- **Running tests:**
+  Tests are executed using [pytest](https://docs.pytest.org/):
   ```bash
   pytest tests
   ```
 
-- **Building documentation**
+- **Building documentation:**
   Documentation is built with [Sphinx](https://www.sphinx-doc.org/).
   A convenience script is provided:
 
@@ -103,7 +103,7 @@ Developers may find the following guidelines useful:
   ./generate_docs.sh
   ```
 
-- **Coding style**
+- **Coding style:**
   The project follows the [Black](https://black.readthedocs.io/en/stable/) code style.
   Please run Black before submitting a pull request:
 
@@ -111,7 +111,7 @@ Developers may find the following guidelines useful:
   black .
   ```
 
-- **Pre-commit hooks**
+- **Pre-commit hooks:**
   [Pre-commit](https://pre-commit.com/) hooks are configured to enforce consistent style automatically.
   To enable them:
 
