### 2025-10-04 9b97f23
update README.rst
