+++ docs/source/api/mdopt.utils.rst
@@ -1,0 +1,15 @@
+mdopt.utils package
+===================
+
+.. automodule:: mdopt.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   mdopt.utils.utils
