+++ docs/source/api/mdopt.optimiser.dephasing_dmrg.rst
@@ -1,0 +1,7 @@
+mdopt.optimiser.dephasing\_dmrg module
+======================================
+
+.. automodule:: mdopt.optimiser.dephasing_dmrg
+   :members:
+   :undoc-members:
+   :show-inheritance:
