+++ docs/source/api/mdopt.contractor.contractor.rst
@@ -1,0 +1,7 @@
+mdopt.contractor.contractor module
+==================================
+
+.. automodule:: mdopt.contractor.contractor
+   :members:
+   :undoc-members:
+   :show-inheritance:
