+++ docs/source/api/mdopt.rst
@@ -1,0 +1,18 @@
+mdopt package
+=============
+
+.. automodule:: mdopt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   mdopt.contractor
+   mdopt.mps
+   mdopt.optimiser
+   mdopt.utils
