+++ docs/source/api/mdopt.utils.utils.rst
@@ -1,0 +1,7 @@
+mdopt.utils.utils module
+========================
+
+.. automodule:: mdopt.utils.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
