+++ docs/source/api/mdopt.optimiser.dmrg.rst
@@ -1,0 +1,7 @@
+mdopt.optimiser.dmrg module
+===========================
+
+.. automodule:: mdopt.optimiser.dmrg
+   :members:
+   :undoc-members:
+   :show-inheritance:
