+++ docs/source/api/mdopt.optimiser.utils.rst
@@ -1,0 +1,7 @@
+mdopt.optimiser.utils module
+============================
+
+.. automodule:: mdopt.optimiser.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
