+++ docs/source/api/mdopt.optimiser.rst
@@ -1,0 +1,17 @@
+mdopt.optimiser package
+=======================
+
+.. automodule:: mdopt.optimiser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   mdopt.optimiser.dephasing_dmrg
+   mdopt.optimiser.dmrg
+   mdopt.optimiser.utils
