+++ docs/source/examples_auto.rst
@@ -1,0 +1,20 @@
+Notebook Examples (auto-generated)
+==================================
+
+The following notebooks were discovered and copied from the repository's
+``examples/`` directory. This page is auto-generated by the docs build script.
+For a curated, tutorial-style set of examples, see :doc:`examples`.
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Notebooks:
+
+   notebooks/classical_ldpc.ipynb
+   notebooks/ground_state.ipynb
+   notebooks/main_component.ipynb
+   notebooks/maxbonddim.ipynb
+   notebooks/mps-rand-circ.ipynb
+   notebooks/quantum_five_qubit.ipynb
+   notebooks/quantum_surface.ipynb
+   notebooks/quantum_three_qubit.ipynb
+   notebooks/shor.ipynb
