+++ docs/source/api/mdopt.mps.rst
@@ -1,0 +1,17 @@
+mdopt.mps package
+=================
+
+.. automodule:: mdopt.mps
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   mdopt.mps.canonical
+   mdopt.mps.explicit
+   mdopt.mps.utils
