+++ docs/source/api/mdopt.contractor.rst
@@ -1,0 +1,15 @@
+mdopt.contractor package
+========================
+
+.. automodule:: mdopt.contractor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   mdopt.contractor.contractor
