+++ docs/source/conf.py
@@ -63,6 +63,10 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 #
 html_theme = "sphinx_rtd_theme"
 nbsphinx_allow_errors = True
+html_theme_options = {
+    "collapse_navigation": True,
+    "navigation_depth": 2,
+}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
