### 2025-10-06 27c87f3
update generate_docs.sh
