+++ generate_docs.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
+set -euo pipefail
 
 PACKAGE_DIR="mdopt"
 DOCS_DIR="docs"
 SPHINX_SOURCE_DIR="docs/source"
 SPHINX_BUILD_DIR="docs/build"
+NOTEBOOKS_DEST_DIR="$SPHINX_SOURCE_DIR/notebooks"
+API_DEST_DIR="$SPHINX_SOURCE_DIR/api"
+EXAMPLES_AUTO_RST="$SPHINX_SOURCE_DIR/examples_auto.rst"
+README_RST="$SPHINX_SOURCE_DIR/README.rst"
 
 # Ensure the script is executed from the project root
 if [ ! -d "$PACKAGE_DIR" ] || [ ! -d "$DOCS_DIR" ]; then
@@ -11,60 +16,73 @@ if [ ! -d "$PACKAGE_DIR" ] || [ ! -d "$DOCS_DIR" ]; then
     exit 1
 fi
 
-# Convert README.md to README.rst
-echo "Converting README.md to README.rst..."
-pandoc "./README.md" -o "$SPHINX_SOURCE_DIR/README.rst"
+echo "==> Preparing Sphinx source tree"
+mkdir -p "$SPHINX_SOURCE_DIR" "$SPHINX_BUILD_DIR" "$NOTEBOOKS_DEST_DIR" "$API_DEST_DIR"
 
-# Delete existing Jupyter Notebooks in the Sphinx source directory
-echo "Deleting existing Jupyter Notebooks in Sphinx source directory..."
-find "$SPHINX_SOURCE_DIR" -name "*.ipynb" -exec rm {} +
+# 1) Convert README.md to README.rst (safe to overwrite)
+echo "==> Converting README.md to $README_RST"
+pandoc "./README.md" -o "$README_RST"
 
-# Copy Jupyter Notebooks to the Sphinx source directory
-# Skip the notebooks whose names start with "tmp" or "plotting"
-for notebook in examples/*/*.ipynb; do
-    notebook_name=$(basename "$notebook")
-    if [[ $notebook_name != tmp* && $notebook_name != plotting* ]]; then
-        cp "$notebook" "$SPHINX_SOURCE_DIR/"
-    fi
-done
-
-# Path to examples.rst
-EXAMPLES_RST="$SPHINX_SOURCE_DIR/examples.rst"
-
-# Check if examples.rst exists. If yes, delete it.
-if [ -f "$EXAMPLES_RST" ]; then
-    echo "examples.rst already exists. Deleting and recreating..."
-    rm "$EXAMPLES_RST"
+# 2) Refresh notebooks in docs/source/notebooks (without touching other .rst files)
+echo "==> Syncing example notebooks into $NOTEBOOKS_DEST_DIR"
+find "$NOTEBOOKS_DEST_DIR" -mindepth 1 -maxdepth 1 -type f -name "*.ipynb" -exec rm -f {} +
+if compgen -G "examples/*/*.ipynb" > /dev/null; then
+    # Skip notebooks starting with tmp or plotting
+    shopt -s nullglob
+    for nb in examples/*/*.ipynb; do
+        base="$(basename "$nb")"
+        if [[ "$base" != tmp* && "$base" != plotting* ]]; then
+            cp "$nb" "$NOTEBOOKS_DEST_DIR/"
+        fi
+    done
+    shopt -u nullglob
+else
+    echo "    (No notebooks found under examples/*/*.ipynb)"
 fi
 
-# Proceed to create (or recreate) examples.rst and fill it
-echo "Creating examples.rst..."
[…]
