### 2025-10-07 quicophy/mdopt#420 @HectorMozo3110
@meandmytram Thanks for your response and for updating everything - the new README looks great.

Thanks for clarifying the Windows support note in the README, as I also observed some discrepancies related to that, particularly with dependency versions between packages like qldpc and numpy.

Before finalizing my review, I’ll run some tests on Linux to verify these points as well. I’ll let you know if I find anything relevant. Regarding the documentation issue, I’ll review it soon and respond accordingly.

Having addressed all points related to this matter, I will proceed to close this issue.
