### 2025-10-07 3b03ccf
better plotting
