+++ pyproject.toml
@@ -38,6 +38,7 @@ ipykernel = "^6.23.1"
 notebook = ">=6.5.3,<8.0.0"
 jupyter = "^1.0.0"
 matplotlib = "^3.7.1"
+pyinstrument = "^5.1.1"
 
 [tool.poetry.group.test]
 optional = true
