### 2025-10-07 30c39f9
marry parallelism and profiling
