+++ examples/decoding/quantum_surface.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 
-lattice_sizes=(5)                               # Array of lattice sizes
-bond_dims=(250)                                 # Array of bond dimensions
-seeds=(10 11 12 13 14 15 16 17 18 19)           # Array of (10) random seeds
-num_experiments=100                             # Runs per each random seed
+lattice_sizes=(3)                               # Array of lattice sizes
+bond_dims=(64)                                  # Array of bond dimensions
+seeds=(100)                                     # Array of random seeds
+num_experiments=10                              # Runs per each random seed
 error_model="Bitflip"                           # The error model
 bias_probs=(1e-1)                               # Array of decoder bias probabilities
 tolerances=(0)                                  # Array of numerical tolerances for the MPS within the decoder
 cuts=(0)                                        # Array of SVD cut-offs for the MPS within the decoder
-num_processes=16                                # The number of processes to use in parallel
+num_processes=1                                 # The number of processes to use in parallel
 silent=false                                    # Whether to suppress the output of the Python script
 
 error_rates=()
@@ -33,7 +33,7 @@ for seed in "${seeds[@]}"; do
                             # Print current configuration
                             echo "Running for lattice size ${lattice_size}, bond dimension ${bond_dim}, error rate ${error_rate}, error model ${error_model}, and seed ${seed}."
                             # Run the Python script
-                            poetry run python examples/decoding/quantum_surface.py --lattice_size ${lattice_size} --bond_dim ${bond_dim} --error_rate ${error_rate} --num_experiments ${num_experiments} --bias_prob ${bias_prob} --error_model "${error_model}" --seed ${seed} --num_processes ${num_processes} --silent ${silent} --tolerance ${tolerance} --cut ${cut}
+                            poetry run python -m pyinstrument -r html -o profile.html examples/decoding/quantum_surface.py --lattice_size ${lattice_size} --bond_dim ${bond_dim} --error_rate ${error_rate} --num_experiments ${num_experiments} --bias_prob ${bias_prob} --error_model "${error_model}" --seed ${seed} --num_processes ${num_processes} --silent ${silent} --tolerance ${tolerance} --cut ${cut}
                         done
                     done
                 done
