### 2025-10-08 quicophy/mdopt#419 @HectorMozo3110
Thank you for the updates, @meandmytram. I have a few observations regarding the documentation.

For the main landing page, it is not necessary to include the entire README content. A concise software description, a brief usage example, and key references would be sufficient and help maintain consistency between the README and the documentation.

What I would recommend is adding these sections to the documentation index and positioning it on the left-hand sidebar rather than within the main page. Including a clear and hierarchical structure with sections such as examples, tests, API reference, and contribution guidelines would improve navigation and make the documentation easier to follow. This would help users and contributors quickly locate relevant information while presenting the software in a more systematic.

A possible structure could include:

- **Getting Started:** Installation instructions, minimal usage example, and workflow overview.
- **Tutorials and Examples:** Jupyter notebooks and scripts demonstrating classical and quantum decoding, tensor-network applications, and representative use cases.
- **Testing and Validation:** Brief explanation of the test suite (e.g., pytest) and instructions for users to run tests directly from the documentation.
- **Contributing:** Incorporation of the current [Code of Conduct](https://github.com/quicophy/mdopt/blob/main/CODE_OF_CONDUCT.md) and contribution guidelines from the README, ensuring they are accessible within the documentation.
- **API Reference:** Direct links to each main module (e.g., `mdopt.mps`, `mdopt.optimiser`, `mdopt.contractor`, etc.).
- **Project Overview and References:** Scientific motivation, context, and links to related publications or external resources.

This list is only an example, and you can adapt it to the current content of the documentation, which already looks good. Still, having a more complete index with clearly defined sections would make it easier to follow. Since the left-hand sidebar is already available, it would be helpful to include all the main sections there for better navigation. I would also recommend adding the `CODE_OF_CONDUCT.md` file, and within that same section, you could include the `pytest` examples that are now presented on the main documentation page. This would give future contributors a clear idea of how to validate and extend the software’s test coverage. Let me know what you think about this suggestion.
