+++ docs/source/index.rst
@@ -1,23 +1,53 @@
-.. include:: README.rst
+Documentation
+===================
 
-.. contents::
-   :local:
-   :depth: 2
+``mdopt`` is a Python package for discrete optimisation in the tensor-network
+language (MPS/MPO). It targets problems such as quantum error-correction decoding with a code-agnostic workflow and is intended to be easily extensible to other applications.
+
+For installation and a minimal working example, see :doc:`getting_started`.
+For a curated overview of notebooks, see :doc:`tutorials`.
 
 .. toctree::
    :maxdepth: 2
-   :caption: Tutorials and notebooks
+   :caption: Getting started
 
+   getting_started
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Tutorials & examples
+
+   tutorials
    examples
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Testing & validation
+
+   testing
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contributing
+
+   contributing
 
 .. toctree::
    :maxdepth: 2
    :caption: API reference
 
+   api_reference
    modules
 
+.. toctree::
+   :maxdepth: 2
+   :caption: Project overview & references
+
+   project_overview
+   README
+
 Indices and tables
-==================
+------------------
 
 * :ref:`genindex`
 * :ref:`modindex`
