+++ docs/source/api_reference.rst
@@ -1,0 +1,14 @@
+API reference
+=============
+
+This section links to the autogenerated API docs.
+
+- :doc:`modules` — full module index (autogenerated by ``sphinx-apidoc``)
+
+Entry points
+------------
+
+- :mod:`mdopt.mps` — MPS data structures and helpers
+- :mod:`mdopt.optimiser` — optimisation via DMRG-like routines
+- :mod:`mdopt.contractor` — tensor-network contraction utilities
+- :mod:`mdopt.utils` — shared tensor utilities
