+++ docs/source/project_overview.rst
@@ -1,0 +1,23 @@
+Project Overview & References
+=============================
+
+Motivation
+----------
+
+``mdopt`` expresses discrete optimisation tasks in the tensor-network (MPS/MPO) formalism.
+This enables controlled approximations (e.g., bond-dimension truncation) and leverages
+well-studied contraction/optimisation techniques for problems such as decoding quantum
+error-correcting codes.
+
+Intended users
+--------------
+
+- Physicists and quantum-information researchers
+- Practitioners exploring tensor-network methods for discrete optimisation
+
+References & related resources
+------------------------------
+
+- Project repository: https://github.com/quicophy/mdopt
+- Documentation (this site): https://mdopt.readthedocs.io/en/latest/
+- Citation details: see the README or project metadata
