+++ docs/source/contributing.rst
@@ -1,0 +1,24 @@
+Contributing
+============
+
+We welcome contributions of all sizes! Before opening a pull request, please:
+
+1. Run the test suite (see :doc:`testing`).
+2. Follow the **Black** code style:
+
+   .. code-block:: bash
+
+      black .
+
+3. Install pre-commit hooks:
+
+   .. code-block:: bash
+
+      pre-commit install
+
+Code of conduct
+---------------
+
+Please read our Code of Conduct:
+
+- `CODE_OF_CONDUCT.md (GitHub) <https://github.com/quicophy/mdopt/blob/main/CODE_OF_CONDUCT.md>`_
