+++ README.md
@@ -1,7 +1,7 @@
 # `mdopt` — Discrete Optimisation in the MPS-MPO Language
 
 <p align="center">
-  <img src="docs/source/logo.png" alt="logo" width="220">
+  <img src="docs/source/logo.png" alt="logo" width="1100">
 </p>
 
 [![codecov](https://codecov.io/gh/quicophy/mdopt/branch/main/graph/badge.svg?token=4G7VWYX0S2)](https://codecov.io/gh/quicophy/mdopt)
