+++ docs/source/index.rst
@@ -7,7 +7,7 @@ Documentation
    :align: center
    :target: https://github.com/quicophy/mdopt
 
-``mdopt`` is a Python package for discrete optimisation in the tensor-network
+mdopt is a Python package for discrete optimisation in the tensor-network
 language (MPS/MPO/DMRG). It targets problems such as quantum error-correction decoding with a code-agnostic workflow and is intended to be easily extensible to other applications.
 
 For installation and a minimal working example, see :doc:`getting_started`.
