### 2025-10-11 836d107
create backend/array.py
