### 2025-10-11 c73be4d
enable BLAS fallback
