### 2025-10-11 5ce1151
adapt to gpu
