+++ mdopt/contractor/contractor.py
@@ -3,15 +3,15 @@ This module contains the MPS-MPO contractor functions.
 """
 
 from typing import Union, List, Tuple, cast
-import numpy as np
 from opt_einsum import contract
 
+from mdopt.backend import array as A
 from mdopt.mps.canonical import CanonicalMPS
 from mdopt.mps.explicit import ExplicitMPS
 from mdopt.utils.utils import split_two_site_tensor
 
 
-def apply_one_site_operator(tensor: np.ndarray, operator: np.ndarray) -> np.ndarray:
+def apply_one_site_operator(tensor, operator):
     """
     Applies a one-site operator to a MPS as follows::
 
@@ -49,26 +49,24 @@ def apply_one_site_operator(tensor: np.ndarray, operator: np.ndarray) -> np.ndar
         raise ValueError(
             f"A valid MPS tensor must have 3 legs while the one given has {tensor.ndim}."
         )
-
-    if len(operator.shape) != 2:
+    if operator.ndim != 2:
         raise ValueError(
-            "A valid one-site operator must have 2 legs"
-            f"while the one given has {len(operator.shape)}."
+            "A valid one-site operator must have 2 legs "
+            f"while the one given has {operator.ndim}."
         )
 
     tensor_updated = contract("ijk, jl -> ilk", tensor, operator, optimize=[(0, 1)])
-
-    return np.asarray(tensor_updated)
+    return A.xp.asarray(tensor_updated)
 
 
 def apply_two_site_unitary(
     lambda_0: list,
-    b_1: np.ndarray,
-    b_2: np.ndarray,
-    unitary: np.ndarray,
+    b_1,
+    b_2,
+    unitary,
     chi_max: int = int(1e4),
     cut: float = float(1e-17),
-) -> Tuple[np.ndarray, np.ndarray]:
+) -> Tuple:
     """
     Applies a two-site unitary operator to a right-canonical MPS as follows::
 
@@ -119,29 +117,29 @@ def apply_two_site_unitary(
 
     """
 
-    if len(b_1.shape) != 3:
+    if b_1.ndim != 3:
         raise ValueError(
-            f"A valid MPS tensor must have 3 legs while the b_1 given has {len(b_1.shape)}."
+            f"A valid MPS tensor must have 3 legs while b_1 has {b_1.ndim}."
+        )
+    if b_2.ndim != 3:
+        raise ValueError(
+            f"A valid MPS tensor must have 3 legs while b_2 has {b_2.ndim}."
+        )
+    if unitary.ndim != 4:
+        raise ValueError(
+            f"A valid two-site operator must have 4 legs while the one given has {unitary.ndim}."
         )
 
-    if len(b_2.shape) != 3:
-        raise ValueError(
-            f"A valid MPS tensor must have 3 legs while the b_2 given has {len(b_1.shape)}."
-        )
-
-    if len(unitary.shape) != 4:
-        raise ValueError(
-            "A valid two-site operator must have 4 legs"
-            f"while the one given has {len(unitary.shape)}."
-        )
-
+    lam = A.xp.asarray(lambda_0)
+    # two-site with lambda_0 on the left bond
     two_site_tensor_with_lambda_0 = contract(
-        "ij, jkl, lmn -> ikmn", np.diag(lambda_0), b_1, b_2, optimize=[(0, 1), (0, 1)]
+        "ij, jkl, lmn -> ikmn", A.xp.diag(lam), b_1, b_2, optimize=[(0, 1), (0, 1)]
     )
     two_site_tensor_with_lambda_0 = contract(
         "ijkl, jkmn -> imnl", two_site_tensor_with_lambda_0, unitary, optimize=[(0, 1)]
     )
[…]
