### 2025-10-12 fc505a6
adapt for gpu
