+++ mdopt/mps/utils.py
@@ -28,12 +28,10 @@ def create_state_vector(num_sites: int, phys_dim: int = int(2)) -> np.ndarray:
     state_vector : np.ndarray
         The resulting state vector.
     """
-
     state_vector = np.random.uniform(
         size=(phys_dim**num_sites)
     ) + 1j * np.random.uniform(size=phys_dim**num_sites)
     state_vector /= np.linalg.norm(state_vector)
-
     return state_vector
 
 
@@ -67,7 +65,6 @@ def find_orth_centre(
         If an :class:`ExplicitMPS` instance is passed as an input.
         They do not have orthogonality centres by definition.
     """
-
     if isinstance(mps, ExplicitMPS):
         raise ValueError(
             "Orthogonality centre is undefined for an Explicit MPS instance."
@@ -87,8 +84,12 @@ def find_orth_centre(
             contract("ijk, ljk -> il", tensor, np.conjugate(tensor), optimize=[(0, 1)])
         )
 
-        identity_left = np.identity(to_be_identity_left.shape[0])
-        identity_right = np.identity(to_be_identity_right.shape[0])
+        identity_left = np.eye(
+            to_be_identity_left.shape[0], dtype=to_be_identity_left.dtype
+        )
+        identity_right = np.eye(
+            to_be_identity_right.shape[0], dtype=to_be_identity_right.dtype
+        )
 
         norm_left = np.linalg.norm(to_be_identity_left - identity_left)
         norm_right = np.linalg.norm(to_be_identity_right - identity_right)
@@ -112,7 +113,6 @@ def find_orth_centre(
         if orth_flags_right == [not flag for flag in orth_flags_left]:
             orth_centres.append(num_sites - 1)
 
-    # Handling a product state.
     if (orth_flags_left == [True] * num_sites) and (
         orth_flags_right == [True] * num_sites
     ):
@@ -146,7 +146,6 @@ def is_canonical(mps: CanonicalMPS, tolerance: float = 1e-12):
         If an :class:`ExplicitMPS` instance is passed as an input.
         They do not have orthogonality centres by definition.
     """
-
     if isinstance(mps, ExplicitMPS):
         raise ValueError(
             "Orthogonality centre is undefined for an Explicit MPS instance."
@@ -162,8 +161,12 @@ def is_canonical(mps: CanonicalMPS, tolerance: float = 1e-12):
             contract("ijk, ljk -> il", tensor, np.conjugate(tensor), optimize=[(0, 1)])
         )
 
-        identity_left = np.identity(to_be_identity_left.shape[0], dtype=float)
-        identity_right = np.identity(to_be_identity_right.shape[0], dtype=float)
+        identity_left = np.eye(
+            to_be_identity_left.shape[0], dtype=to_be_identity_left.dtype
+        )
+        identity_right = np.eye(
+            to_be_identity_right.shape[0], dtype=to_be_identity_right.dtype
+        )
 
         flags_left.append(
             np.isclose(
@@ -180,9 +183,7 @@ def is_canonical(mps: CanonicalMPS, tolerance: float = 1e-12):
         return True
 
     orth_centres = find_orth_centre(mps)
-    if_canonical = len(orth_centres) == 1
-
-    return if_canonical
+    return len(orth_centres) == 1
 
 
 def inner_product(
@@ -208,7 +209,6 @@ def inner_product(
     ValueError
         If the Matrix Product States are of different length.
     """
-
     if len(mps_1) != len(mps_2):
         raise ValueError(
             f"The number of sites in the first MPS is {len(mps_1)} while "
@@ -217,7 +217,6 @@ def inner_product(
[…]
