### 2025-10-12 6089282
adapt for gpu
