+++ mdopt/optimiser/utils.py
@@ -3,7 +3,6 @@ This module contains different combinatorial optimisation utilities.
 
 First, we define the tensors which represent logical operation.
 We use the following tensors: IDENTITY, COPY, XOR, SWAP.
-See the notes for additional information.
 
 According to our convention, each tensor has legs (vL, vR, pU, pD),
 where v stands for "virtual", p -- for "physical",
@@ -22,43 +21,31 @@ from mdopt.utils.utils import mpo_to_matrix
 from mdopt.contractor.contractor import mps_mpo_contract
 
 
-IDENTITY = np.eye(2).reshape((1, 1, 2, 2))
+IDENTITY = np.eye(2, dtype=float).reshape((1, 1, 2, 2))
 
 COPY_LEFT = np.fromfunction(
-    lambda i, j, k, l: np.eye(2)[j, l],
-    (1, 2, 2, 2),
-    dtype=int,
-)
+    lambda i, j, k, l: np.eye(2)[j, l], (1, 2, 2, 2), dtype=int
+).astype(float, copy=False)
 
 XOR_BULK = np.fromfunction(
-    lambda i, j, k, l: (i ^ j ^ k ^ 1) * np.eye(2)[k, l],
-    (2, 2, 2, 2),
-    dtype=int,
-)
+    lambda i, j, k, l: (i ^ j ^ k ^ 1) * np.eye(2)[k, l], (2, 2, 2, 2), dtype=int
+).astype(float, copy=False)
 
 XOR_LEFT = np.fromfunction(
-    lambda i, j, k, l: np.eye(2)[j, k] * np.eye(2)[k, l],
-    (1, 2, 2, 2),
-    dtype=int,
-)
+    lambda i, j, k, l: np.eye(2)[j, k] * np.eye(2)[k, l], (1, 2, 2, 2), dtype=int
+).astype(float, copy=False)
 
 XOR_RIGHT = np.fromfunction(
-    lambda i, j, k, l: np.eye(2)[i, k] * np.eye(2)[k, l],
-    (2, 1, 2, 2),
-    dtype=int,
-)
+    lambda i, j, k, l: np.eye(2)[i, k] * np.eye(2)[k, l], (2, 1, 2, 2), dtype=int
+).astype(float, copy=False)
 
 SWAP = np.fromfunction(
-    lambda i, j, k, l: np.eye(2)[i, j] * np.eye(2)[k, l],
-    (2, 2, 2, 2),
-    dtype=int,
-)
+    lambda i, j, k, l: np.eye(2)[i, j] * np.eye(2)[k, l], (2, 2, 2, 2), dtype=int
+).astype(float, copy=False)
 
 
 def parity(bitstring: str, indices: list[int]) -> int:
-    """
-    Returns the parity of the bits at the given indices in the bitstring.
-    """
+    """Returns the parity of the bits at the given indices in the bitstring."""
     return sum(int(bitstring[i]) for i in indices) % 2
 
 
@@ -102,11 +89,9 @@ def random_constraints(num_bits: int, constraint_size: int, rng: Generator) -> d
         )
 
     indices = sorted(rng.choice(num_bits, size=constraint_size, replace=False))
-
     xor_left_sites = [indices[0]]
     xor_right_sites = [indices[-1]]
     xor_bulk_sites = indices[1:-1]
-
     swap_sites = [
         i for i in range(indices[0] + 1, indices[-1]) if i not in xor_bulk_sites
     ]
@@ -147,17 +132,13 @@ class ConstraintString:
     """
 
     def __init__(self, constraints: List[np.ndarray], sites: List[List[int]]) -> None:
-        self.constraints = [
-            np.array(constraint, dtype=float) for constraint in constraints
-        ]
[…]

[…]
