### 2025-10-12 9569f8d
adapt for gpu
