+++ mdopt/utils/utils.py
@@ -1,10 +1,29 @@
-"""This module contains miscellaneous utilities."""
+"""This module contains different tensor utilities."""
 
 from typing import Tuple, Optional, List, cast
 from itertools import chain
 import numpy as np
 import scipy
 from opt_einsum import contract
+
+# --- Backend shim: prefer your GPU/array backend if available, else NumPy ---
+try:
+    # expected to export a NumPy-like API (e.g., NumPy or CuPy)
+    from mdopt.backend import array as xp  # type: ignore
+except Exception:
+    import numpy as xp  # type: ignore
+
+
+def _to_numpy(a):
+    """Convert backend arrays (e.g., CuPy) to NumPy without copying if possible."""
+    try:
+        import cupy as cp  # type: ignore
+
+        if isinstance(a, cp.ndarray):
+            return cp.asnumpy(a)
+    except Exception:
+        pass
+    return np.asarray(a)
 
 
 def svd(
@@ -46,69 +65,77 @@ def svd(
     ValueError
         If the input matrix does not have 2 dimensions.
     """
-
     if mat.ndim != 2:
         raise ValueError(
             f"A valid matrix must have 2 dimensions while the one given has {mat.ndim}."
         )
 
-    svd_methods = [
-        lambda: np.linalg.svd(
-            mat, full_matrices=False, compute_uv=True, hermitian=False
-        ),
-        lambda: scipy.linalg.svd(
-            mat,
-            full_matrices=False,
-            compute_uv=True,
-            check_finite=True,
-            lapack_driver="gesdd",
-        ),
-        lambda: scipy.linalg.svd(
-            mat,
-            full_matrices=False,
-            compute_uv=True,
-            check_finite=True,
-            lapack_driver="gesvd",
-        ),
-        lambda: scipy.linalg.svd(
-            mat + np.eye(mat.shape[0], mat.shape[1]) * 1e-12,
-            full_matrices=False,
-            compute_uv=True,
-            check_finite=True,
-            lapack_driver="gesvd",
-        ),
-    ]
+    # Try backend SVD first (GPU-friendly), then fall back to SciPy variants.
     last_exception: Optional[Exception] = None
-    for method in svd_methods:
+    u_l = s = v_h = None  # type: ignore
+    a = xp.asarray(mat)
+    for attempt in ("xp", "gesdd", "gesvd", "jitter"):
         try:
-            u_l, singular_values, v_r = method()
+            if attempt == "xp":
+                u_l, s, v_h = xp.linalg.svd(a, full_matrices=False)  # returns U, S, Vh
+            elif attempt == "gesdd":
+                u_l, s, v_h = scipy.linalg.svd(
+                    _to_numpy(a),
+                    full_matrices=False,
+                    compute_uv=True,
+                    check_finite=True,
+                    lapack_driver="gesdd",
+                )
+            elif attempt == "gesvd":
+                u_l, s, v_h = scipy.linalg.svd(
+                    _to_numpy(a),
+                    full_matrices=False,
+                    compute_uv=True,
+                    check_finite=True,
+                    lapack_driver="gesvd",
+                )
+            else:
+                an = _to_numpy(a)
+                u_l, s, v_h = scipy.linalg.svd(
+                    an + np.eye(an.shape[0], an.shape[1]) * 1e-12,
+                    full_matrices=False,
+                    compute_uv=True,
+                    check_finite=True,
+                    lapack_driver="gesvd",
+                )
[…]
