### 2025-10-12 1d628b6
adapt for gpu
