### 2025-10-12 a2e4f04
mypy patch
