### 2025-10-13 8e0480d
adapt for gpu
