+++ mdopt/contractor/contractor.py
@@ -2,8 +2,9 @@
 This module contains the MPS-MPO contractor functions.
 """
 
-import numpy as np
 from typing import Union, List, Tuple, cast
+
+import numpy as np
 from opt_einsum import contract
 
 from mdopt.backend import array as A
@@ -45,7 +46,7 @@ def apply_one_site_operator(tensor, operator):
     ValueError
         If the operator tensor is not two-dimensional.
     """
-
+    backend = "cupy" if A.GPU else "numpy"
     if tensor.ndim != 3:
         raise ValueError(
             f"A valid MPS tensor must have 3 legs while the one given has {tensor.ndim}."
@@ -56,7 +57,9 @@ def apply_one_site_operator(tensor, operator):
             f"while the one given has {operator.ndim}."
         )
 
-    tensor_updated = contract("ijk, jl -> ilk", tensor, operator, optimize=[(0, 1)])
+    tensor_updated = contract(
+        "ijk, jl -> ilk", tensor, operator, optimize=[(0, 1)], backend=backend
+    )
     return A.to_device(np.asarray(tensor_updated))
 
 
@@ -117,7 +120,7 @@ def apply_two_site_unitary(
         If the operator tensor is not four-dimensional.
 
     """
-
+    backend = "cupy" if A.GPU else "numpy"
     if b_1.ndim != 3:
         raise ValueError(
             f"A valid MPS tensor must have 3 legs while b_1 has {b_1.ndim}."
@@ -131,23 +134,27 @@ def apply_two_site_unitary(
             f"A valid two-site operator must have 4 legs while the one given has {unitary.ndim}."
         )
 
-    lam = np.asarray(lambda_0)
-    # two-site with lambda_0 on the left bond
-    two_site_tensor_with_lambda_0 = contract(
-        "ij, jkl, lmn -> ikmn", np.diag(lam), b_1, b_2, optimize=[(0, 1), (0, 1)]
+    # device-consistent lambda and broadcast onto left virtual leg of b_1
+    lam = A.to_device(np.asarray(lambda_0)) if A.GPU else np.asarray(lambda_0)
+    b1_scaled = b_1 * (lam[:, None, None])
+
+    # with lambda_0
+    t_with = contract(
+        "ijk, klm -> ijlm", b1_scaled, b_2, optimize=[(0, 1)], backend=backend
     )
-    two_site_tensor_with_lambda_0 = contract(
-        "ijkl, jkmn -> imnl", two_site_tensor_with_lambda_0, unitary, optimize=[(0, 1)]
+    t_with = contract(
+        "ijkl, jkmn -> imnl", t_with, unitary, optimize=[(0, 1)], backend=backend
     )
 
-    # two-site without lambda_0 (for back-substitution)
-    two_site_tensor_wo_lambda_0 = contract("ijk, klm", b_1, b_2, optimize=[(0, 1)])
-    two_site_tensor_wo_lambda_0 = contract(
-        "ijkl, jkmn -> imnl", two_site_tensor_wo_lambda_0, unitary, optimize=[(0, 1)]
+    # without lambda_0 (for back-substitution)
+    t_wo = contract("ijk, klm -> ijlm", b_1, b_2, optimize=[(0, 1)], backend=backend)
+    t_wo = contract(
+        "ijkl, jkmn -> imnl", t_wo, unitary, optimize=[(0, 1)], backend=backend
     )
 
+    # split and back-substitute
     _, _, b_2_updated, _ = split_two_site_tensor(
-        two_site_tensor_with_lambda_0,
+        t_with,
         chi_max=chi_max,
         cut=cut,
         renormalise=False,
@@ -155,12 +162,16 @@ def apply_two_site_unitary(
[…]
