### 2025-10-13 0fa0b0d
mypy patch
