### 2025-10-13 9d682e6
Bump numpy from 1.26.4 to 2.2.6 (#402)
* Bump numpy from 1.26.4 to 2.2.6

Bumps [numpy](https://github.com/numpy/numpy) from 1.26.4 to 2.2.6.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/main/doc/RELEASE_WALKTHROUGH.rst)
- [Commits](https://github.com/numpy/numpy/compare/v1.26.4...v2.2.6)

---
updated-dependencies:
- dependency-name: numpy
  dependency-version: 2.2.6
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>

* [pre-commit.ci] auto fixes
[…]
