+++ .pre-commit-config.yaml
@@ -39,6 +39,6 @@ repos:
           exclude: '(^examples|^docs)/.*'
 
   - repo: https://github.com/python-poetry/poetry
-    rev: 1.8.0
+    rev: 2.2.1
     hooks:
       - id: poetry-check
