### 2025-10-18 #9125 @HectorMozo3110
> 👋 [@meandmytram](https://github.com/meandmytram)
>
> These are my initial observations based on the JOSS review checklist. As part of this, I opened a few issues that you’ve probably already noticed, mainly to make observations and suggestions for improvement.
>
> We can use those threads to develop the discussion point by point. As we go covering them, I’ll check them off here. I will also update this list if I open additional issues, and once we’ve gone through all of them I’ll be able to close my review.
>
> * [x]  [#419 Suggestions for Documentation (JOSS review)](https://github.com/quicophy/mdopt/issues/419)
> * [x]  [#420 Observations on the README (JOSS review)](https://github.com/quicophy/mdopt/issues/420)
> * [x]  [#423 Docs reference added in paper (JOSS review)](https://github.com/quicophy/mdopt/pull/423)

👋 @meandmytram

Thanks for the updates -  I’ve gone through my checklist, and everything looks great. All previous issues have been fully addressed.

I’ve been replicating the examples and testing the code, including the **5-qubit** and **Shor’s 9-qubit** decoding cases, as well as Monte-Carlo runs for logical error rates. Everything executed correctly, results were consistent, and the tensor-network behavior looked solid throughout.

The only small thing I noticed is the **NumPy–Numba compatibility**, but it seems you’re already updating those dependencies, so that shouldn’t be a problem.

With this, I’m happy to say that I recommend this submission for acceptance. @danielskatz
