+++ requirements.txt
@@ -558,9 +558,9 @@ osqp==1.0.4 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:bae336207b52bbab6d663e7fb77beaffc53279e91a14e3db76ad2726b0e935d6 \
     --hash=sha256:c2a2ac69feb009ac43f80f33dfe5b32baaccefd22c6fcb9176b48f755506f120 \
     --hash=sha256:ddd7fc52e01c602f3482878b334eed1b366a9940b2053059968d3d47688de9ef
-packaging==24.2 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:09abb1bccd265c01f4a3aa3f7a7db064b36514d2cba19a2f694fe6150451a759 \
-    --hash=sha256:c228a6dc5e932d346bc5739379109d49e8853dd8223571c7c5b55260edc0b97f
+packaging==25.0 ; python_version >= "3.10" and python_version < "4.0" \
+    --hash=sha256:29572ef2b1f17581046b3a2227d5c611fb25ec70ca1ba8554b24b0e69331a484 \
+    --hash=sha256:d443872c98d677bf60f6a1f2f8c1cb748e8fe762d2bf9d3148b5599295b0fc4f
 pandas==2.3.1 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:025e92411c16cbe5bb2a4abc99732a6b132f439b8aab23a59fa593eb00704232 \
     --hash=sha256:09e3b1587f0f3b0913e21e8b32c3119174551deb4a4eba4a89bc7377947977e7 \
