### 2025-10-24 e84e866
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
