+++ docs/source/examples_auto.rst
@@ -2,7 +2,7 @@ Notebook Examples (auto-generated)
 ==================================
 
 The following notebooks were discovered and copied from the repository's
-examples directory. This page is auto-generated by the docs build script.
+``examples/`` directory. This page is auto-generated by the docs build script.
 For a curated, tutorial-style set of examples, see :doc:`examples`.
 
 .. toctree::
