+++ pyproject.toml
@@ -25,7 +25,8 @@ dependencies = [
     "qldpc>=0.1.2,<0.2.0",
     "sympy>=1.14.0,<2.0.0",
     "seaborn>=0.13.2,<0.14.0",
-    "ldpc>=2.3.6,<3.0.0"
+    "ldpc>=2.3.6,<3.0.0",
+    "pydata-sphinx-theme (>=0.16.1,<0.17.0)"
 ]
 
 [tool.poetry.group.dev]
@@ -57,7 +58,6 @@ optional = true
 [tool.poetry.group.docs.dependencies]
 matplotlib = "^3.7.0"
 Sphinx = ">=6.2.1,<9.0.0"
-sphinx-rtd-theme = ">=2,<4"
 myst-parser = ">=2,<5"
 sphinxcontrib-napoleon = "^0.7"
 nbsphinx = "^0.9.2"
