+++ docs/source/requirements.txt
@@ -1,7 +1,7 @@
 sphinx==5.3.0
 sphinxcontrib-napoleon==0.7
 myst-parser==0.18.1
-pydata_sphinx_theme
+pydata_sphinx_theme==0.16.1
 ipykernel
 nbsphinx
 ipython
