### 2025-10-28 cd2bcc4
remove logo background
