+++ docs/source/conf.py
@@ -72,6 +72,7 @@ html_theme_options = {
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+html_css_files = ["css/custom.css"]
 
 # This will delete the annoying 'bases: object' line from the docs.
 
