### 2025-10-29 35c84b2
Bump cvxpy from 1.6.6 to 1.7.3
Bumps [cvxpy](https://github.com/cvxpy/cvxpy) from 1.6.6 to 1.7.3.
- [Release notes](https://github.com/cvxpy/cvxpy/releases)
- [Commits](https://github.com/cvxpy/cvxpy/compare/v1.6.6...v1.7.3)

---
updated-dependencies:
- dependency-name: cvxpy
  dependency-version: 1.7.3
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
