+++ requirements.txt
@@ -124,33 +124,33 @@ contourpy==1.2.1 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:ef5adb9a3b1d0c645ff694f9bca7702ec2c70f4d734f9922ea34de02294fdf72 \
     --hash=sha256:f32c38afb74bd98ce26de7cc74a67b40afb7b05aae7b42924ea990d51e4dac02 \
     --hash=sha256:fe0ccca550bb8e5abc22f530ec0466136379c01321fd94f30a22231e8a48d985
-cvxpy==1.6.6 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:060e09cdc0cb7044ee8b4fd736168ecda86edcf64af3f39ab579f59a36336ba3 \
-    --hash=sha256:091bf83fb4b7d58ea73380325f654239be6e7f556f2546c1cb767c41bf6815e3 \
-    --hash=sha256:0ad3abd3dabe5ae5d4d88b53d62d13d53c3515dbbd4ffb3348df91eef6b23171 \
-    --hash=sha256:0bc83872e7054434c9c242a1a154daacae4c57f2818c7edfefcbf69a0f628748 \
-    --hash=sha256:101fb9b433cc2e334d3596ad611fcc3e2fa5d484bb0c65eda9ee2d7213de05b2 \
-    --hash=sha256:200f969a171e7b7f6d682e51c9c595e7fec7014de551355a915aad16e9a824b2 \
-    --hash=sha256:220711c4cef001804613eccdbe1d104ba2be705b47ba1f8dc073ebc542bc5ca5 \
-    --hash=sha256:23eb8dad835c425630a5d6b249c8685bd5d9038ad564cf6317e9ae259046b654 \
-    --hash=sha256:3b3e2b5db609434608c7f535ab483b055c3012ca02bb97b1cc76a97ef0ea9fef \
-    --hash=sha256:3d011ca756ea13632a739431a59b3a0d17e36e8dab280059763c09c3d615662f \
-    --hash=sha256:5a23a18a1b88b008996b3cb696b8b64c315c44bd875b449658784f8a0b70aa02 \
-    --hash=sha256:5d2aecceeac9f5b9297c26ea0080a8de0da94f2660103c6b2c996f6900fb9dab \
-    --hash=sha256:5d8a7a567823ea43b5e24f390bb298ebd905cfb3570b4feeed0bcb80815f1400 \
-    --hash=sha256:73b82adb9a32ac75d98b52b0520dfcea846e5d5fe7d84c539b101043d51d532b \
-    --hash=sha256:7790b15365411778acbe25d23368f4e2776f28f44c7cbb6059101072e40fc028 \
-    --hash=sha256:845bb27ad6d7c8069e1fa769b638a0f07dfb4cbaa2f1e92dbc3c4ef5eb6cdeab \
-    --hash=sha256:863535ba8d89806a8cfcd5cb7939aef7ec7c45a7109780af72ceda6090410887 \
-    --hash=sha256:9bf2c311becaac48ea566692c53dba9bd39f0c4dd10534945191e39fd398e7ee \
-    --hash=sha256:9d7aa7ecf5409459aaa5f39ec4889a472314d5790f478c093b6ebe733b4e4bd3 \
-    --hash=sha256:b424f2416b2d8935628e1291e97d532ec34ae046246fe9d2d2d69115ff1ba701 \
-    --hash=sha256:b461bd6cb9dd975ca5bf1ac86a3602036158eca7cdff8d824e8a7c30efe43f66 \
[…]
