### 2025-10-29 8edcbad
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
