+++ requirements.txt
@@ -124,33 +124,33 @@ contourpy==1.2.1 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:ef5adb9a3b1d0c645ff694f9bca7702ec2c70f4d734f9922ea34de02294fdf72 \
     --hash=sha256:f32c38afb74bd98ce26de7cc74a67b40afb7b05aae7b42924ea990d51e4dac02 \
     --hash=sha256:fe0ccca550bb8e5abc22f530ec0466136379c01321fd94f30a22231e8a48d985
-cvxpy==1.7.3 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:052db7e5485d3cbe6b0249c4cc0fec898cdfafcb8d2490149a21bd9a7d00a82f \
-    --hash=sha256:1496a6993fd0c1d6b16af78a075318c0ad34d99c48efac26fe07e61aa17bd296 \
-    --hash=sha256:241d364f5962a1d68c4ae8393480766a09326e5771e2286d33a948e1976cbe70 \
-    --hash=sha256:2adc901fcff104b54b75f002cce5de442dec88aaeb9732d273446c34a21cdcf0 \
-    --hash=sha256:2d8b2213296478478d267537681f96ea7d9941d5bb1fa61717797f9fabd3b747 \
-    --hash=sha256:44c021a3732092f687e7d693524e4cf90aff449b40eeeb1c4e871fac0b83c2f4 \
-    --hash=sha256:68168ff87b0fd2157861a31fcfb538cf5d68c069a31016be89a70a2e5c2860ca \
-    --hash=sha256:753b14a630d28784769fe2b696f240095df4100e49bae08cd330fc72c746e4a5 \
-    --hash=sha256:7a97e4e3eff4e9f8151058fe7687c7eecfb44b6a1ae83834195b701088ce29e7 \
-    --hash=sha256:7bd145daf239b8a235895f36ff0611ff6fff2cad844290a8f1c6df7055b9cb98 \
-    --hash=sha256:7ee60aabd035ae20e0948ee23cf98b1c99009b6b75780d73b811c9c4232c97ad \
-    --hash=sha256:939e76a84e1a243f32e83d0dbeede4274f827a002106575e0cd6515b51f140ce \
-    --hash=sha256:9801a9c8b0c9aee206fba6d25d4d1d959a5c25a45a9b678c9b8ce625e51e0d67 \
-    --hash=sha256:990815f0fdf956cdf7e8a11d64c90bad9080635a7f21884165a50e42c64924e9 \
-    --hash=sha256:a214e3f0ccb31b0584cb44f7a11c652b28c1fc255c84d33ac8d84dcb216657d7 \
-    --hash=sha256:abe973a4a03c7beb775b017786f6cd9b3138f0d0c0f01b3a49b75b3314c0e7d8 \
-    --hash=sha256:b7743b261b92e12aef5a7ed9593314e4ceb6cba2c897b21adab70ef02d2ca54c \
-    --hash=sha256:ba0ebbe7e8e84bd7c87fcc225a6ad724caa7af2d515ccd227ec37cfb884bbcae \
-    --hash=sha256:c06fc9904ea5de9c0cb3993e2bd2a98bd6b32c0ba79b7b813b0855feceef7aa6 \
-    --hash=sha256:cd043d40febf77fa14ad516d85b760002b9c235e2193bd85cb094b57fd521c69 \
-    --hash=sha256:cfed5afe3d53fc7835ea8cc9e281d0bbf3b4edcc625e8a1162c42cb807b8830d \
[…]
