### 2025-10-30 #9125 @WhiteSymmetry
### Subject: Issues with PyPI Release of `mdopt` Package

The current version of the `mdopt` package on PyPI (https://pypi.org/project/mdopt/) contains code that is outdated compared to the Git repository, leading to runtime errors.

There are two primary issues with the code distributed via PyPI:

1.  **Incorrect Import:** The example code incorrectly imports the `apply_constraints` function from `examples.decoding.decoding`. The correct import path is `from mdopt.optimiser.utils import apply_constraints`.

2.  **Missing/Deprecated Parameter:** The example code uses a `result_to_explicit` keyword argument when calling the `apply_constraints` function. This parameter is documented but is not present in the function's signature in the released code, causing a `TypeError`.

While the `README.md` file on the Git repository may have been updated with a corrected example, the source code bundled in the PyPI release has not been updated accordingly. Consequently, a `pip install mdopt` fetches an outdated and non-functional version of the example code.

**Proposed Solution:**
A new version should be published to PyPI. This will ensure that the `pip` installation delivers the same corrected and functional code that is currently available in the Git repository.

**(For reference, the corrected code snippet is provided below):**

```python
# Corrected imports
from mdopt.optimiser.utils import apply_constraints # Correct source
# from examples.decoding.decoding import apply_constraints # Incorrect, removed

# Corrected function call (the 'renormalise' parameter is removed)
logging.info("Applying bitflip bias to the perturbed codeword state.")
        perturbed_codeword_state = apply_bitflip_bias(
            mps=perturbed_codeword_state,
            sites_to_bias="All",
            prob_bias_list=PROB_BIAS,
            # renormalise=True, # This parameter has been removed
        )

# Corrected function call (the 'result_to_explicit' parameter is removed)
perturbed_codeword_state = apply_constraints(
    perturbed_codeword_state,
    code_constraint_sites,
    tensors,
    chi_max=CHI_MAX,
    renormalise=True,
    # result_to_explicit=False, # This parameter has been removed
    strategy="Optimised",
    silent=False,
)
```
