+++ docs/source/conf.py
@@ -26,7 +26,7 @@ copyright = str(datetime.date.today().year) + ", Aleksandr Berezutskii"
 author = "Aleksandr Berezutskii"
 
 # The full version, including alpha/beta/rc tags
-release = "1.0.0"
+release = "1.0.1"
 
 
 # -- General configuration ---------------------------------------------------
