### 2025-11-03 #9125 @meandmytram
@WhiteSymmetry can you elaborate more on your comment please? It is not entirely clear for me what you mean by "convergence behaviour" and especially "accuracy" in the context of decoding? Just FYI, we do not provide any performance claims of the software at this point.
