### 2025-11-03 7e152a5
add module docstring
