### 2025-11-03 858f94d
exception catch
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
