+++ pyproject.toml
@@ -26,7 +26,6 @@ dependencies = [
     "sympy>=1.14.0,<2.0.0",
     "seaborn>=0.13.2,<0.14.0",
     "ldpc>=2.3.6,<3.0.0",
-    "pydata-sphinx-theme (>=0.16.1,<0.17.0)"
 ]
 
 [tool.poetry.group.dev]
@@ -63,6 +62,7 @@ sphinxcontrib-napoleon = "^0.7"
 nbsphinx = "^0.9.2"
 pygments = "^2.17.2"
 ipython = "^8.22.1"
+pydata-sphinx-theme = ">=0.16.1,<0.17.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
