### 2025-11-03 5d8cc31
enhance __getattr__ fallback
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
