### 2025-11-03 f278745
use list comprehension
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
