### 2025-11-03 9226715
use list comprehension
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
