### 2025-11-03 0978dbf
Backend options (#424)
* create backend/array.py

* enable BLAS fallback

* adapt the codebase for gpu

* update environments

* update docstrings

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* update environment

* catch several typos

* extend description

* new theme


* update mdopt version

* remove logo background

* add module docstring

* add gpu example

* better plotting

* update docs

* exception catch

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* exception catch

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.co
[…]
