+++ mdopt/optimiser/dephasing_dmrg.py
@@ -43,6 +43,9 @@ class EffectiveDensityOperator(scipy.sparse.linalg.LinearOperator):
     The diagram displaying the contraction can be found in the supplementary notes.
     """
 
+    # Single compiled einsum route kept for readability/reuse
+    _EINSUM = "ustw, ailu, ifj, jhk, bef, cgh, lom, mpn, eso, gtp, dknw -> abcd"
+
     def __init__(
         self,
         left_environment: np.ndarray,
@@ -50,68 +53,46 @@ class EffectiveDensityOperator(scipy.sparse.linalg.LinearOperator):
         mps_target_2: np.ndarray,
         right_environment: np.ndarray,
     ) -> None:
-        """
-        Initialise an effective dephased density operator tensor network.
-
-        Parameters
-        ----------
-        left_environment : np.ndarray
-            The left environment for the effective dephased density operator.
-        mps_target_1 : np.ndarray
-            The left target matrix product state tensor.
-        mps_target_2 : np.ndarray
-            The right target matrix product state tensor.
-        right_environment : np.ndarray
-            The right environment for the effective dephased density operator.
-
-        Raises
-        ------
-        ValueError
-            If the left environment tensor is not four-dimensional.
-        ValueError
-            If the first target MPS tensor is not three-dimensional.
-        ValueError
-            If the second target MPS tensor is not three-dimensional.
-        ValueError
-            If the right environment tensor is not four-dimensional.
-        """
-        if len(left_environment.shape) != 4:
+        if left_environment.ndim != 4:
             raise ValueError(
-                "A valid left environment tensor must have 4 legs"
-                f"while the one given has {len(left_environment.shape)}."
+                "A valid left environment tensor must have 4 legs "
+                f"while the one given has {left_environment.ndim}."
             )
-        if len(mps_target_1.shape) != 3:
+        if mps_target_1.ndim != 3:
             raise ValueError(
-                "A valid target MPS tensor must have 3 legs"
-                f"while the one given has {len(mps_target_1.shape)}."
+                "A valid target MPS tensor must have 3 legs "
+                f"while the one given has {mps_target_1.ndim}."
             )
-        if len(mps_target_2.shape) != 3:
+        if mps_target_2.ndim != 3:
             raise ValueError(
-                "A valid target MPS tensor must have 3 legs"
-                f"while the one given has {len(mps_target_1.shape)}."
+                "A valid target MPS tensor must have 3 legs "
+                f"while the one given has {mps_target_2.ndim}."
             )
-        if len(right_environment.shape) != 4:
+        if right_environment.ndim != 4:
             raise ValueError(
-                "A valid right environment tensor must have 4 legs"
-                f"while the one given has {len(right_environment.shape)}."
+                "A valid right environment tensor must have 4 legs "
+                f"while the one given has {right_environment.ndim}."
             )
 
         self.left_environment = left_environment
         self.right_environment = right_environment
         self.mps_target_1 = mps_target_1
         self.mps_target_2 = mps_target_2
-        chi_1, chi_2 = (
-            left_environment.shape[3],
-            right_environment.shape[3],
-        )
-        d_1, d_2 = (
-            mps_target_1.shape[1],
-            mps_target_2.shape[1],
-        )
+
[…]
