+++ mdopt/optimiser/dmrg.py
@@ -37,6 +37,8 @@ class EffectiveOperator(scipy.sparse.linalg.LinearOperator):
     are ordered as follows: ``(uL/uR, vL/vR, dL/dR)`` which means "(up, virtual, down)".
     """
 
+    _EINSUM = "ijkl, mni, nopj, oqrk, sql -> mprs"
+
     def __init__(
         self,
         left_environment: np.ndarray,
@@ -44,84 +46,54 @@ class EffectiveOperator(scipy.sparse.linalg.LinearOperator):
         mpo_tensor_right: np.ndarray,
         right_environment: np.ndarray,
     ) -> None:
-        """
-        Initialises an effective operator tensor network.
-
-        Parameters
-        ----------
-        left_environment : np.ndarray
-            The left environment for the effective operator.
-        mpo_tensor_left : np.ndarray
-            The left MPO tensor.
-        mpo_tensor_right : np.ndarray
-            The right MPO tensor.
-        right_environment : np.ndarray
-            The right environment for the effective operator.
-        """
-
-        if len(left_environment.shape) != 3:
+        if left_environment.ndim != 3:
             raise ValueError(
-                "A valid left environment tensor must have 3 legs"
-                f"while the one given has {len(left_environment.shape)}."
+                "A valid left environment tensor must have 3 legs "
+                f"while the one given has {left_environment.ndim}."
             )
-        if len(mpo_tensor_left.shape) != 4:
+        if mpo_tensor_left.ndim != 4:
             raise ValueError(
-                "A valid mpo left tensor must have 4 legs"
-                f"while the one given has {len(mpo_tensor_left.shape)}."
+                "A valid mpo left tensor must have 4 legs "
+                f"while the one given has {mpo_tensor_left.ndim}."
             )
-        if len(mpo_tensor_right.shape) != 4:
+        if mpo_tensor_right.ndim != 4:
             raise ValueError(
-                "A valid mpo right tensor must have 4 legs"
-                f"while the one given has {len(mpo_tensor_right.shape)}."
+                "A valid mpo right tensor must have 4 legs "
+                f"while the one given has {mpo_tensor_right.ndim}."
             )
-        if len(right_environment.shape) != 3:
+        if right_environment.ndim != 3:
             raise ValueError(
-                "A valid right environment tensor must have 3 legs"
-                f"while the one given has {len(right_environment.shape)}."
+                "A valid right environment tensor must have 3 legs "
+                f"while the one given has {right_environment.ndim}."
             )
 
         self.left_environment = left_environment
         self.right_environment = right_environment
         self.mpo_tensor_left = mpo_tensor_left
         self.mpo_tensor_right = mpo_tensor_right
-        chi_1, chi_2 = (
-            left_environment.shape[2],
-            right_environment.shape[2],
-        )
-        d_1, d_2 = (
-            mpo_tensor_left.shape[3],
-            mpo_tensor_right.shape[3],
-        )
+
+        chi_1 = left_environment.shape[2]
+        chi_2 = right_environment.shape[2]
+        d_1 = mpo_tensor_left.shape[3]
+        d_2 = mpo_tensor_right.shape[3]
+
         self.x_shape = (chi_1, d_1, d_2, chi_2)
         self.shape = (chi_1 * d_1 * d_2 * chi_2, chi_1 * d_1 * d_2 * chi_2)
         self.dtype = mpo_tensor_left.dtype
         super().__init__(shape=self.shape, dtype=self.dtype)
 
[…]
