### 2025-11-04 #9125 @danielskatz
@WhiteSymmetry - Can you please affirm that you are performing this review manually, and not using AI?
