### 2025-11-07 33f8b9a
🐍 Add Python classifiers
