### 2025-11-07 17eb407
📄 Add license information via PEP 639
