+++ pyproject.toml
@@ -10,6 +10,8 @@ maintainers = [
 ]
 readme = "README.md"
 keywords = [ "tensor networks", "quantum error correction", "discrete optimisation", "MPS", "MPO", "python", "decoding" ]
+license = "MIT"
+license-files = ["LICENSE.md"]
 requires-python = ">=3.10,<4.0"
 
 dependencies = [
