### 2025-11-07 daddb35
👷 Update RtD configuration
