### 2025-11-07 quicophy/mdopt#435 @burgholzer
Limited CI testing
### What is your issue?

The current CI setup only tests on Ubuntu with Python 3.10.
Given how cheap CI resources are on GitHub for public projects (i.e., free), it should be a no brainer to add more elaborate testing to the repository.
Tests like these would have also revealed that the library currently is not compatible with Python 3.14.

This is part of https://github.com/openjournals/joss-reviews/issues/9125

## What should be added?

Comprehensive CI testing on all supported Python (3.10-3.13) and OS versions (`ubuntu-24.04`, `macos-15`, `windows-2025`).
I'd personally also recommend trying to test on `ubuntu-24.04-arm`, `macos-15-intel`, as well as `windows-11-arm`, if your dependencies permit it.
