### 2025-11-07 quicophy/mdopt#436 @burgholzer
Missing CD workflow
### What is your issue?

The repository is currently missing a CD workflow for automatic releases, which makes me assume that the release process is handled manually at the moment.

This is part of https://github.com/openjournals/joss-reviews/issues/9125

## What should be added

A GitHub workflow triggered on GitHub releases that builds the source distribution as well as the wheel and uploads them to PyPI (preferably using trusted publishing).
Some information on trusted publishing https://docs.pypi.org/trusted-publishers/

You could take https://github.com/munich-quantum-toolkit/bench/blob/main/.github/workflows/cd.yml as an inspiration, which internally uses
- https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-packaging-sdist.yml
- https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-packaging-wheel-build.yml
