+++ pyproject.toml
@@ -10,7 +10,7 @@ maintainers = [
 ]
 readme = "README.md"
 keywords = [ "tensor networks", "quantum error correction", "discrete optimisation", "MPS", "MPO", "python", "decoding" ]
-requires-python = ">=3.10,<4.0"
+requires-python = ">=3.10,<3.14"
 
 dependencies = [
     "scipy>=1.9.2,<2.0.0",
