### 2025-11-07 122b7e3
add CI badge
