+++ .github/workflows/ci.yml
@@ -39,14 +39,14 @@ jobs:
 
   # Canary for next CPython (marked non-blocking)
   py314:
-    name: canary (ubuntu-24.04, py3.14)
+    name: canary (ubuntu-24.04, py3.13)
     runs-on: ubuntu-24.04
     continue-on-error: true
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.14"
+          python-version: "3.13"
           cache: pip
           cache-dependency-path: pyproject.toml
       - run: |
