### 2025-11-07 quicophy/mdopt#437 @meandmytram
Thanks for the suggestion Lukas, however I am putting this issue on hold at the moment because pylint still does its job really well and scanning the whole codebase takes seconds on my laptop.
