+++ .github/workflows/ci.yml
@@ -63,8 +63,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        # arm64 Linux; Intel macOS image
-        os: [ubuntu-24.04-arm, macos-15-intel]
+        # Intel macOS image
+        os: [macos-15-intel]
         python: ["3.12"]
     steps:
       - uses: actions/checkout@v4
