### 2025-11-07 38db191
🏷️ Mark Python package as typed (#430)
