### 2025-11-07 a4c066a
Update .readthedocs.yml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
