+++ .readthedocs.yml
@@ -3,7 +3,7 @@ version: 2
 build:
   os: ubuntu-24.04
   tools:
-    python: "3.13"
+    python: "3.10"
 
 sphinx:
    configuration: docs/source/conf.py
