### 2025-11-07 a58d0b2
📄 Add license information via PEP 639 (#432)
