### 2025-11-07 799a00e
🐍 Add Python classifiers (#431)
* 🏷️ Mark Python package as typed

* 🐍 Add Python classifiers

---------

Co-authored-by: Aleksandr (Sasha) Berezutskii <aleksandr.berezutskii@usherbrooke.ca>
