+++ pyproject.toml
@@ -14,6 +14,12 @@ license = "MIT"
 license-files = ["LICENSE.md"]
 requires-python = ">=3.10,<4.0"
 classifiers = [
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Typing :: Typed",
 ]
 
