### 2025-11-07 80fbbf3
Addressing the JOSS review (#438)
This PR bumps the package version from 1.0.1 to 1.1.1 and introduces comprehensive CI/CD workflows for automated testing and publishing.

Version update from 1.0.1 to 1.1.1
Addition of CI workflow for multi-platform and multi-Python version testing
Addition of CD workflow for automated PyPI publishing via GitHub releases

* restrict python version

* multiplatform ci

* add CI badge

* restrict python version

* update ci.yml

* create cd.yml

* update name

* bump version
