### 2025-11-08 fa1c5c5
minor changes in paper.md
