### 2025-11-08 #9125 @danielskatz
@meandmytram - I'm suggesting a few minor changes in https://github.com/quicophy/mdopt/pull/440 - please merge this, or let me know what you disagree with, then we can proceed.
