+++ .github/workflows/cd.yml
@@ -30,7 +30,7 @@ jobs:
         run: |
           python -m pip install -U twine
           python -m twine check dist/*
-      - uses: actions/upload-artifact@v4
+      - uses: actions/upload-artifact@v5
         with:
           name: dist
           path: dist/*
