+++ .github/workflows/cd.yml
@@ -44,7 +44,7 @@ jobs:
       name: testpypi
       url: https://test.pypi.org/p/mdopt
     steps:
-      - uses: actions/download-artifact@v4
+      - uses: actions/download-artifact@v6
         with:
           name: dist
           path: dist
@@ -63,7 +63,7 @@ jobs:
       name: pypi
       url: https://pypi.org/p/mdopt
     steps:
-      - uses: actions/download-artifact@v4
+      - uses: actions/download-artifact@v6
         with:
           name: dist
           path: dist
